🔧Fix issues with flag cache (/ploot reload will now correctly update the default flag) 🔧Providers will now try to find the item before the entity (dropping ``arrow`` spawned arrow entities instead of item) 🔧Fix the internal calculator messing up with big numbers (makes the discord flag and others that use big numbers work again) 🆕/ploot now says the plugin version 🆕"DISPLAY" as a color option - it will use the display color of the item 💻Combined "LootProviderManager" and "ItemProviderManager" into "Providers"

+++ Plugin/src/main/java/net/seyarada/pandeloot/utils/MathUtils.java
@@ -56,7 +56,7 @@ public class MathUtils {
         return dx * dx + dz * dz;
     }
 
-    public static double eval(final String str) {
+    public static long eval(final String str) {
         return new Object() {
             int pos = -1, ch;
 
@@ -73,9 +73,9 @@ public class MathUtils {
                 return false;
             }
 
-            double parse() {
+            long parse() {
                 nextChar();
-                double x = parseExpression();
+                long x = parseExpression();
                 if (pos < str.length()) throw new RuntimeException("Unexpected: " + (char)ch);
                 return x;
             }
@@ -86,8 +86,8 @@ public class MathUtils {
             // factor = `+` factor | `-` factor | `(` expression `)`
             //        | number | functionName factor | factor `^` factor
 
-            double parseExpression() {
-                double x = parseTerm();
+            long parseExpression() {
+                long x = parseTerm();
                 for (;;) {
                     if      (eat('+')) x += parseTerm(); // addition
                     else if (eat('-')) x -= parseTerm(); // subtraction
@@ -95,8 +95,8 @@ public class MathUtils {
                 }
             }
 
-            double parseTerm() {
-                double x = parseFactor();
+            long parseTerm() {
+                long x = parseFactor();
                 for (;;) {
                     if      (eat('*')) x *= parseFactor(); // multiplication
                     else if (eat('/')) x /= parseFactor(); // division
@@ -104,34 +104,34 @@ public class MathUtils {
                 }
             }
 
-            double parseFactor() {
+            long parseFactor() {
                 if (eat('+')) return parseFactor(); // unary plus
                 if (eat('-')) return -parseFactor(); // unary minus
 
-                double x;
+                long x;
                 int startPos = this.pos;
                 if (eat('(')) { // parentheses
                     x = parseExpression();
                     eat(')');
                 } else if ((ch >= '0' && ch <= '9') || ch == '.') { // numbers
                     while ((ch >= '0' && ch <= '9') || ch == '.') nextChar();
-                    x = Double.parseDouble(str.substring(startPos, this.pos));
+                    x = Long.parseLong(str.substring(startPos, this.pos));
                 } else if (ch >= 'a' && ch <= 'z') { // functions
                     while (ch >= 'a' && ch <= 'z') nextChar();
                     String func = str.substring(startPos, this.pos);
                     x = parseFactor();
                     x = switch (func) {
-                        case "sqrt" -> Math.sqrt(x);
-                        case "sin" -> Math.sin(Math.toRadians(x));
-                        case "cos" -> Math.cos(Math.toRadians(x));
-                        case "tan" -> Math.tan(Math.toRadians(x));
+                        case "sqrt" -> (long) Math.sqrt(x);
+                        case "sin" -> (long) Math.sin(Math.toRadians(x));
+                        case "cos" -> (long) Math.cos(Math.toRadians(x));
+                        case "tan" -> (long) Math.tan(Math.toRadians(x));
                         default -> throw new RuntimeException("Unknown function: " + func);
                     };
                 } else {
                     throw new RuntimeException("Unexpected: " + (char)ch);
                 }
 
-                if (eat('^')) x = Math.pow(x, parseFactor()); // exponentiation
+                if (eat('^')) x = (long) Math.pow(x, parseFactor()); // exponentiation
 
                 return x;
             }
